refactor: slowly refactoring initial guess
